requirements.txt
pandas>=2.0
pyarrow>=14.0
pyyaml>=6.0
numpy>=1.24
python-dateutil>=2.8
openpyxl>=3.1
# Postgres loader (step 07)
sqlalchemy>=2.0
psycopg[binary]>=3.1
